Removed web only button, replaced with web/native button
