+++ src/Button/index.js
@@ -1,5 +1,3 @@
 import Button from './Button';
-import ButtonNative from './ButtonNative';
 
-export { ButtonNative };
 export default Button;
