+++ src/native/index.js
@@ -1,7 +1,7 @@
 import Text, { H1, H2, H3, H4, H5, P, Quote, Ingress, Caption, InfoText, ErrorText, ListText } from '../Typography';
 import { NewsFeedItemNative, NewsFeedNative } from '../NewsFeed';
 import { Tabs, Tab } from '../Tabs';
-import { ButtonNative } from '../Button';
+import Button from '../Button';
 import Touchable from '../Touchable';
 
 export {
@@ -20,7 +20,7 @@ export {
   ListText,
   NewsFeedItemNative as NewsFeedItem,
   NewsFeedNative as NewsFeed,
-  ButtonNative as Button,
+  Button,
   Touchable,
   Tabs,
   Tab
